Work on the new ACPICA integration
I am experiencing a problem with the integretation, the extracted IRQ
Routings seems to be all zero's, and this causes multiple registrations
of irq 0. Another problem is that the size of the kernel has skyrocketed
from 180kb to 440kb.

+++ kernel/acpica/utilities/utstrsuppt.c
@@ -265,7 +265,7 @@ AcpiUtConvertDecimalString (
     {
         /* Character must be ASCII 0-9, otherwise terminate with no error */
 
-        if (!isdigit (*String))
+        if (!isdigit ((int)*String))
         {
            break;
         }
@@ -321,7 +321,7 @@ AcpiUtConvertHexString (
     {
         /* Must be ASCII A-F, a-f, or 0-9, otherwise terminate with no error */
 
-        if (!isxdigit (*String))
+        if (!isxdigit ((int)*String))
         {
             break;
         }
